Redefined Components and CSS
