--- src/Components/ListLostItems.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import "../styles/LostItems.css";
-import ListItem from "./LostItem";
-
-const ListLostItems = () => {
-  return (
-    <div className="container">
-      <ListItem />
-      <ListItem />
-      <ListItem />
-      <ListItem />
-      <ListItem />
-      <ListItem />
-    </div>
-  );
-};
-
-export default ListLostItems;
