+++ src/Components/Lost/ListItem.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import "../../styles/Lost/Items.css";
+import Item from "./Item";
+
+const ListLostItems = () => {
+  return (
+    <div className="container">
+      <Item />
+      <Item />
+      <Item />
+      <Item />
+    </div>
+  );
+};
+
+export default ListLostItems;
