+++ src/styles/Lost/Card.css
@@ -7,6 +7,8 @@
 
 .lostCard > svg {
   position: absolute;
+  border: 1px solid red;
+  width: 100%;
   bottom: 0;
   z-index: 1;
   transform-origin: bottom;
