+++ src/styles/Lost/Items.css
@@ -1,0 +1,33 @@
+.container {
+  /* border: 1px solid brown; */
+  display: grid;
+  place-items: center;
+  grid-template-columns: repeat(3, 1fr);
+  padding: 20px 20px;
+}
+
+/* .container > div {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  border: 1px solid brown;
+} */
+
+.itemContainer {
+  overflow: hidden;
+  width: 400px;
+  margin: 10px;
+  border-radius: 15px;
+  border: 1px solid blue;
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
+    0 2px 4px -1px rgba(0, 0, 0, 0.06);
+  display: flex;
+  align-items: flex-start;
+  gap: 10px;
+}
+
+.innerContainer {
+  border: 1px solid green;
+  height: 100%;
+  margin-top: 20px;
+}
