+++ src/styles/Lost/NavBar.css
@@ -3,8 +3,7 @@
   justify-content: space-between;
   align-items: center;
   padding: 20px 50px;
-  box-shadow: 0 10px 20px -1px rgba(0, 0, 0, 0.1),
-    0 2px 4px -1px rgba(0, 0, 0, 0.06);
+  border-bottom: 1px solid gray;
 }
 
 .navBar > div:nth-child(2) {
@@ -19,4 +18,8 @@
 
 .cursor-pointer {
   cursor: pointer;
+}
+.navItem {
+  cursor: pointer;
+  color: #019aff;
 }
