--- src/styles/LostItems.css
@@ -1,13 +1,0 @@
-.container {
-  padding: 40px;
-  display: flex;
-  flex-direction: row;
-}
-
-.itemContainer {
-  margin: 10px;
-  border-radius: 15px;
-  border: none;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
-    0 2px 4px -1px rgba(0, 0, 0, 0.06);
-}
